refactor: 💡 Move LinkService tests into their own file
